test(fluent-dom): Move elem() from index.js to util.js
