+++ fluent-dom/test/overlay_functional_children_test.js
@@ -1,6 +1,6 @@
 import assert from 'assert';
 import translateElement from '../esm/overlay.js';
-import {elem} from './index';
+import {elem} from './util.js';
 
 suite('Child without name', function() {
   test('in source', function() {
