+++ fluent-dom/test/overlay_test.js
@@ -1,6 +1,6 @@
 import assert from 'assert';
 import translateElement from '../esm/overlay.js';
-import {elem} from './index';
+import {elem} from './util.js';
 
 suite('Applying translations', function() {
   test('Skipping sanitization for the title element', function() {
