+++ fluent-dom/test/overlay_text_children_test.js
@@ -1,6 +1,6 @@
 import assert from 'assert';
 import translateElement from '../esm/overlay.js';
-import {elem} from './index';
+import {elem} from './util.js';
 
 suite('Text-semantic argument elements', function() {
   test('without data-l10n-name', function() {
