+++ fluent-dom/test/util.js
@@ -1,0 +1,7 @@
+export function elem(name) {
+  return function(str) {
+    const element = document.createElement(name);
+    element.innerHTML = str;
+    return element;
+  }
+}
